Move JobType to be inside of ServerInfo
