+++ xrpl4j-model/src/main/java/org/xrpl/xrpl4j/model/client/serverinfo/ServerInfoLoad.java
@@ -9,9 +9,9 @@ package org.xrpl.xrpl4j.model.client.serverinfo;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -25,6 +25,7 @@ import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
 import com.google.common.primitives.UnsignedInteger;
 import org.immutables.value.Value;
+import org.xrpl.xrpl4j.model.client.serverinfo.ServerInfo.JobType;
 
 import java.util.List;
 
@@ -46,8 +47,8 @@ public interface ServerInfoLoad {
   }
 
   /**
-   * (Admin only) Information about the rate of different types of jobs the server is doing and how much time
-   * it spends on each.
+   * (Admin only) Information about the rate of different types of jobs the server is doing and how much time it spends
+   * on each.
    *
    * @return A {@link List} of {@link JobType}s.
    */
