+++ xrpl4j-model/src/test/java/org/xrpl/xrpl4j/model/client/serverinfo/JobTypeTest.java
@@ -5,6 +5,7 @@ import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
 import com.google.common.primitives.UnsignedInteger;
 import org.junit.jupiter.api.Test;
+import org.xrpl.xrpl4j.model.client.serverinfo.ServerInfo.JobType;
 
 /**
  * Unit tests for {@link JobType}.
