+++ xrpl4j-model/src/test/java/org/xrpl/xrpl4j/model/client/serverinfo/ReportingModeServerInfoTest.java
@@ -8,6 +8,7 @@ import com.google.common.primitives.UnsignedLong;
 import org.junit.jupiter.api.Test;
 import org.xrpl.xrpl4j.model.AbstractJsonTest;
 import org.xrpl.xrpl4j.model.client.common.LedgerIndex;
+import org.xrpl.xrpl4j.model.client.serverinfo.ServerInfo.JobType;
 import org.xrpl.xrpl4j.model.transactions.Hash256;
 import org.xrpl.xrpl4j.model.transactions.XrpCurrencyAmount;
