+++ xrpl4j-model/src/test/java/org/xrpl/xrpl4j/model/client/serverinfo/ServerInfoResultTest.java
@@ -19,6 +19,9 @@ import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.Locale;
 
+/**
+ * Unit tests for {@link ServerInfoResult}.
+ */
 public class ServerInfoResultTest extends AbstractJsonTest {
 
   @Test
@@ -75,7 +78,7 @@ public class ServerInfoResultTest extends AbstractJsonTest {
       .time(ZonedDateTime.parse("2021-Mar-30 15:37:51.486384 UTC",
         DateTimeFormatter.ofPattern("yyyy-MMM-dd HH:mm:ss.SSSSSS z", Locale.US)).withZoneSameLocal(ZoneId.of("UTC")))
       .upTime(UnsignedLong.valueOf(2274704))
-      .validatedLedger(ServerInfoLedger.builder()
+      .validatedLedger(ServerInfoValidatedLedger.builder()
         .age(UnsignedInteger.valueOf(4))
         .hash(Hash256.of("E5A958048D98D4EFEEDD2BC3F36D23893BBC1D9354CB3E739068D2DFDE3D1AA3"))
         .reserveBaseXrp(XrpCurrencyAmount.ofDrops(20100000))
